Modernize plugin build (#24)
1) Upgrade Java
Java 7 is end-of-life since quite some time. Therefore raising the required
Java version to 8. In addition raising the required Jenkins version to the
2.60.x line which was the first one that supported Java 8.

2) Uprade Gradle & JPI plugin
In version 0.36.0 of the jpi plugin bundling the classes has been changed.
Instead of putting them into the HPI file under WEB-INF/classes/, they are
now put into a JAR file under WEB-INF/lib/. This is required for Jenkins
JAR file prefetching & caching to work properly (see JENKINS-15120).

3) Update dependencies
Either upgrading them to stay up-to-day or removing some. With the
upgrade of Gradle and the JPI plugin transitive dependencies are
handled much better. This removes the need to declare them explicitly.